refactor: solve issues after review
