+++ src/components/dictionary/Dictionary.tsx
@@ -2,8 +2,8 @@
 import axios from 'axios';
 import React, { useEffect, useState } from 'react';
 import Product from '../words/Product';
-import { IProduct } from '../../interfaces/interfaces';
-import { COLORS } from '../constants';
+import { IProduct } from '../../types';
+import { COLORS } from '../../constants';
 import './dictionary.scss';
 
 export default function Dictionary() {
